✨ Advent of `part_lib`. Roughly usable.
A system to stick Node3Ds together by the means of `PartConnector` objects.
This is done using `PartControl` objects, which are passed a Node3D to act on
upon instantiation. The connectors are gleaned from the Node3D's skeleton if
available, or the fallback skeleton the `PartConnector` generates for itself if
the Node3D doesn't have one (it looks for a node named "Skeleton3D", which
should be there if the object (in Blender) is parented to an armature (assuming
gltf/glb export)).

`PartControl` is designed as a base class which is either fully usable out of
the box for simple scenarios with no more than two bones or to be sub-classed
to support other use cases. Importantly, it provides
`default_receiving_connector` and `default_docking_connector`. If the fallback
skeleton is used they references the same connector.

For sub-classing, there are `._init_...` methods which can be overriden without
having to re-noodle `._init`.

Notably, connector bones are modeled by `Node3DTailedBone`, which means the
system expects each bone to have a counterpart suffixed by `_Tail` by default.
The vector between a bone and its tail is to be understood as the direction
it is facing, with the tail supposed to be the "away" point, as in the point
the vector is pointing towards.