✨ Advent of `city_part_lib` (roughly usable), and stub of `city_parts` (empty).
`BasicBuildingPartControl` is a sub-class of `PartControl` from part_lib, and
introduces two connectors that could serve as a basis for every city part
expected to work with the part system: A bottom left front and bottom right
front connector - based on matching bones, of course.

A sub-class of this is `BasicFacadePartControl`, which introduces a top left
front connector, of course with the expectation of a matching bone on the
skeleton.

A skeleton used for a city part with this system should have the following
bones:
- `Bottom_Left_Front` and `Bottom_Left_Front_Tail`.
- `Bottom_Right_Front` and `Bottom_Right_Front_Tail`.
- and, if it's a facade, also `Top_Left_Front` and `Top_Left_Front_Tail`.

There's also `PartTransformer`, which is concerned with things like changing
the size of a part according to some other metric, like columns in a facade
grid, for example. Currently, the function `respan_3_columns` makes use of it.

A very basic, experimental example of this can be looked at in
(*)`blender_house_assembly_mess.gd`.