✨ Advent of `city_geo_funcs`. Roughly usable.
When messing around with vertices, some functionality is best plopped in a
function, so it's simple to re-use but also for people to just copy and use in
other projects. The "city_" prefix means that these functions have some sort of
bias towards use with city stuff, or at least things related to buildings.

An example of this is `shear_line`, which provides the core functionality of
`.shear_vertices` of `ALine` of mesh_lib. The city bias comes in with the
y-axis getting special treatment.