✨ Advent of `mesh_lib` and `mesh_debug_lib`. Roughly usable.
Mostly classes helpful for building meshes based on godot's ArrayMesh.
They're organized in a class hierarchy which builds on a common base class
`ASegment`, which is a wrapper around an ArrayMesh.ARRAY_VERTEX array, which can
be obtained by calling `.get_array_vertex` (which, in this case, returns
`._array_vertex`). The `A` prefix in these classes stands for "Array", as in
"ArrayMesh", as opposed to hypothetical `AM` classes, which would stand for
"ArrayMesh".

Any class that builds on `ASegment` can override that method and do what's
necessary there for its functionality.

`ATransformableSegment` is a particularly notable sub-class, as it introduces
transformability using `Transform3D`, making it easier to build segments out of
other segments. Notably it keeps an untransformed `ASegment` as a member,
`.untransformed`, which it references whenever its values are (re-)calculated.
It also introduces an important method into the class hierarchy, `.apply_all`,
which re-calculates the internal state of the object, in this case
`._array_vertex`. It may be overriden by sub-clasess as needed.

This also introduces the concept of explicitly calling `.apply_...` methods
after making alterations to segment objects, which makes it possible to have
segment classes which can be altered without frivolous re-calculations before
all the necessary changes for an appropriate re-calculation are done.

`AVertexTrackingSegment` introduces `.vertices: Array[Vertex]`, whereas every
`Vertex` object maps to one or multiple `Vector3` of the vertex array. They act
as delegates of sorts, knowing both an `.untransformed_segment` and a
`.transformed_segment`, with the latter intended to hold a reference to the
segment they belong to and the former its `.untransformed` segment. They're also
explicitly permitted to make direct alterations to `._array_vertex` of both
segments.

This enables the convenient reference and alteration of vertices that overlap
on the same position, but are, for all higher-level purposes, the same vertex.

Then come the segment classes for primitives which are sub-classes of
`AVertexTrackingSegment`, such as `AQuad` or `ALine`.

Also a sub-class of it is `AMultiSegment`, which serves as a base class for
segment classes designed to combine segment objects into more complex segments,
such as `AHorizontallyFoldedTriangle`.

Also included is mesh_debug_lib, providing `ADebugOverlay`, which can be used to
visualize vertices and their values in `ArrayMesh.ARRAY_VERTEX`-arrays.