🐛 Pseudo-fixed camera not working for some scenes.
Currently, the project has defined actions corresponding to `PlayerWorldInterface`, but without keybindings configured. In this case, it won't configure keybindings at runtime, because its decision to do that is based on the existence of an action, not a keybinding.

Until this is handled properly, the affected scenes now have their `PlayerWorldInterface` configured to override existing actions.