🚧 Advent of the `raw_export` Blender addon.
Currently prints JSON for the selected object to the terminal, with vertices, uvs, indices, enumerated materials with a list of images found in their node graph and a "material index" per face (the key of this list is "material indices") which corresponds to the index in the eumerated materials. There's also "coord_texture.png" (hereby released under the terms of CC0), a debugging texture which helps visualize the test Cube's sides as well as a shell script that makes it easy to start the included blend-file with the addon loaded & ready to go (it comes with lots of variables and .env configurability, which should make it easy to accommodate various situations). Also, the README now contains the section "Licensing", clarifying the new license situation now that there's a blender addon in the repo. In the wake of this, "LICENSE" got renamed to "LICENSE-MIT". Added a CC0 section while at it.