✨ Advent of `json_serializable`, a `global_lib` that provides a small virtual base class interface with a `to_json` and a `to_dict` method, whereas particularly the latter is meant for overriding in sub-classes.
`to_json`, per default, calls it to provide a serialized version of the object. If the object should be represented as a list, the idea would be to add a to_list method and override `to_json` to call that instead. Of course, that would not be part of the virtual base class interface then. It's currently intended for the raw_export data classes related to JSON (de)serialization, for which a dict-based approach is sensible.