🍱 `coord_texture.png` is easier to import in  `assets/parts/textures`, so copied it there.
The problem is that `utils/raw_export/blender_addon/dev` is zoned off by .gdignore. The solution might be to use the normal FileAccess API to load the file.

Mind to update that copy when updating the texture, and to add that to the documentation for raw_export once we have any (if the workaround is still in place by then, of course).