✨ Sanity info when exporting the test cube from Blender.
Currently prints verts/uvs grouped by cube side to the console, with a `faces_match` field which shows `True` if the Tris on that side belong together and `False` if they don't.

Specifically, `faces_match` is `True` if the vertices of the two tris belonging to a cube side are all negative or positive on the same axis.

This rests on the assumption that the test cube is centered at Blender's world origin, of course.

There's also an `axis` and `axis_sign` field for each tri, which shows which side of the cube the tri is on. Mind that the axis letters are lower case here to be more consistent and potentially compatible with how the axis are usually referred to in code, unlike in the texture where they're upper case for improved readability.