✨ One-liner import of RawExport meshes is now possible and works with the test cube.
Advent of the global_lib `RawExport4MeshLib` and `STris`, a new kind of class in mesh_lib: A `Surface` (sub-)class. I'm not a 100% sure about the whole concept yet, but at the moment, it's a class with a `get_instance_3d` method. The idea is to keep non-vertex-ish information on segment classes, and wrap segments in surface classes, which take care of things like material assignments, which are sorta face-ish. In order to keep up with segment mutations, a companion system will probably be needed for segments, so `SegmentMutator` will also mutate registered companions. That could be used for metadata, for example, at least the face-ish, edge-ish, or even the irregular kind.