✨ Proper panel for the RawExport Blender addon and wrapper operators for single object export, collection export and exporting all eligible objects.
Every object can now be given a filename and collections can be given a sub-dir via the new panel. There's also a settings category where an overall (base) directory can be configured. The File Export category then allows export of the selected object, all objects in the selected collection (and sub-collections) or all objects, for as long as they are eligible for export (which they are if they have a filename).

The path to write to for each object is the combination of the configured (base) directory, any sub-dirs found in the parent collection hierarchy of an object and its filename.

Caveat: This has not been tested with multiple scenes, and the settings are currently attached to `context.scene`.