💩 Advent of `AFlushingMultiSegment`, an `AMultiSegment` base class designed to clear the source of the return value of `get_segments` after `apply_all` got called.
This is a cheap bandaid covering for a larger design problem with `AMultiSegment` in general: Since it's become clear that applying everything except the actual transform to .untransformed is the way to go, sourcing everything from `.get_segments` for `AMultiSegment` classes is an unintuitive oddity that really rubs the wrong way with `AModifiableSegment`.

This new class provides multi-segment logic that meshes better with the current design goals, but it being a sub-class of `AMultiSegment` is rather awkward. Refactoring the two classes into three, a base class and two sibling classes, might be the better approach.