🚧 Adding and transforming and object to the window which isn't exported from the Blender scene with the same origin as the other parts, and also doesn't perfectly fit the geometry. The lighting isn't quite right, but other than that it works.
The original problem was with the behaviour of `AMultiSegment`, as described in commit 37cf256df12b32e2e437a0474268b2aeb41f4bdf. In this case, the problem was that the transform didn't show up when `decor_crown_surface_tris` was added, since `.add` added from `.get_segments`, which, of course, didn't have the transform, because that's never applied to `.base_tris`, which happens to be the source for `.get_segments` in this case.

The first attempt to solve this was to simply use `.get_array_*` instead of `.get_segments` in `.add`, but then it became apparent that `.add` had only worked before because it'd unintentionally used completely unmodified (as in `.apply_modifiers`) data because of the above described behaviour of `.get_segment` of `AMultiSegment`, which, of course, doesn't apply just to transform, but the whole other shebang as well, like the modifier stack.

The result was a very brittle system, that, if the segments being added together wouldn't get prepared beforehand to be just at the right stage of their planned modifications before getting added together, you'd be pretty much set to get unexpected results (or, even worse, perhaps it might even have been possible to get correct looking results at first, and incorrect results with different input later).

Situations like these were bound to keep happening, which is why commit 37cf256df12b32e2e437a0474268b2aeb41f4bdf introduced `AFlushingMultiSegment`, on which the current solution is based.