🚧 Preparations for `LayoutTool` and some refactoring of `StreetTool`.
Much of the code of `StreetTool` is now in a new `CurveTool` class which now serves as a base class for `StreetTool` and `LayoutTool`. Some further refactoring will be required to fully distill the common base for such tools, particularly the "click for direction, click to plop point, rinse and repeat"-bit, which is quite opinionated in favor of `StreetTool`.

It's also not clear yet whether the current use case for `LayoutTool` does lend itself to being curve based, but a separation of this bit of code will probably prove useful for future curve based tools as well.