🚧 There is now a `Layout` class in the `city_builder` lib, and the layout tool class has been split into a polygon and a curve one.
The reason for the curve/polygon split is that for non-curved facades, we don't need to take the curve route. This would enable more performant visualization, but it might also cause more work down the line, since not all layout tool code will be based on curves. Then again, considering this early might prevent future headaches in that regard, keeping things more flexible, and that's probably more important for citygamesystems.