🏗️ Advent of `ToolLibMapAgent`.
First stab at an abstraction layer decoupling tools from the map. Designed for sub-classing to accommodate different map architectures. Doesn't abstract a whole lot right now, though, particularly with the `get_map_node` method, but this is as far as I'm willing to go at the moment, lest it might become too rigid for prototyping.

It does serve to define an interface for tool-map interaction, however, and effectively puts an end to the type-coupling we've had so far.