🏗️ Architectural work for the city demo and snapping for the layout tool (to its own corner points, not yet to other layouts).
Advent of various concepts: Positioner, WorldObject, LifeCycler (unsure about the naming) and ColliderManager (will probably go away).

A Positioner is a node that can be added to a MapAgent, which has a `get_position` function which uses a positioner to return a potentially altered position (used for snapping). This might be expanded to allow multiple positioners, which could be used by UI to allow for configurable snapping depending on the positioners attached to the tool's MapAgent. Other things, like tools, could also be written to directly use positioners. Positioners are also used in WorldObject (see below).

WorldObject classes provide a basis for things like houses or streets, and currently have some collision related stuff, including a function that returns a MultiPositioner (which can be type-erased to a Positioner) that is meant to be overriden in a subclass, where object-specific positioners can be added to it (e.g. for vertices, curves or domain specific ones).

LifeCycler classes are, sort-of, (de)constructor factories for WorldObject classes. An instance of such a class is responsibe to provide (de)construction for world objects of the WorldObject class it's responsible for. This approach allows to have project-specific (de)construction logic for world objects that interacts with runtime elements, such as statically typed access to nodes in the scene graph, like the `collision_manager` in thea `CityGameLifeCycler` base class, which can only be initialized during instantiation, since the collider manager is a node in the scene tree instead of an Autoload.

The ColliderManager maps world object colliders to world objects. This was an attempt to establish a system that supports collider-to-world-object matching for objects that aren't subclassed from WorldObject. However, such objects are probably better wrapped in a world object that is implemented as an adapter for them, making the ColliderManager quite moot. It's committed for archival purposes, and to prevent further refactoring in this already way-too-large commit.

Also, the layout tool still contains a bunch of street tool stuff for testing.