✨ Advent of `CachingViewportMousePositionGetter`.
This node caches the mouse position of the viewport it's been configured for for the duration of a frame, so the expensive `Viewport.get_mouse_position()` method only has to be called once per frame for that viewport. Of course, you have to use the node's `.position` property in order for that to work.

The currently planned use case for this is going to be the following: Getting the mouse position for raycasting, potentially from different parts of the codebase, which would all incur the cost of `Viewport.get_mouse_position()` individually without caching.

On my machine (Ryzen 5800X3D), each subsequent cached read saves about 60 usecs or so.