🚧 `Mouse3DRay` raycasting now happens in a swappable component.
`Mouse3DRayRayCaster` is named that way to make it clear that it's a component class of `Mouse3DRay`, and, as with other "redundant" names like that throughout the project, to mitigate potential namespace issues.

Also, switching to `Node` for now, as we might not actually need it to be a Node3D (we'll see).