✨ The street tool's live preview now uses `StreetMesh`.
This is also the advent of the new `SignalBuffer` addon. The motivation for it
is to compensate for the problem with `waiting_for_map_points`: The wait makes
sure the street curve isn't getting altered whilst the map raycaster does its
work. Any mouse input received during the wait would be discarded, however,
making for horrible UX. Assuming that the wait is, generally, very short, we're
using `SingleFirstSignalBuffer` here to store the first mouse click during
the wait. The buffer is flushed in `_on_result_map_points`, which simulates
the mouse click by sending the buffered signal that got "intercepted" during
the wait to the street_tool.gd.