🏗️ It's now possible to get a `WorldObject` from a collider.
This either depends on the new collider sub-classes being used, or the colliders having a metadata string, which will have to be set somewhere by something (e.g. a `LifeCycler`).

Adds a new global to `CityGameGlobals`, `MetaNames.WORLD_OBJECT_ON_COLLIDERS`, which is intended to be used with the new `WorldObject.get_from_collider_or_null` function.
Since I'm having a mind of pushing `WorldObject` into a base library, and there's no CityGameSystemsGlobals, this is where it has to go - even though it would be more convenient to have it all baked away right under the hood of `WorldObject` whilst still making the string easy to change. I'm not quite sure that'd be the way to go, though, since I'm still trying to keep the base of citygamesystems a set of "dumb" libraries instead of configurable framework pieces. In any case, the current architecture mandates it be scoped in with the demo, not the base framework (for now?).