🚧 Start of a `StreetSegment` based approach with a collider.
The current idea is that instances of `StreetSegmentWorldObject` represent parts of the street network, with the visualization being its own thing (meshes, shaders, ...).
The street tool is now also using the life cycler approach.

GeoFoo also got a new `vector3` object serving as a namespace for functions designed to "extend" Vector3.