🚧 [Curve2dDebugVisualizer] Implemented curve sampling based visualization.
The old code now lives in its own function, and the updating function has been refactored to call another updating function - that makes it possible to extend the tool one day to make it configurable as to which method to use. Might be useful for debugging Godot's curve sampling by comparing the results of the sampling with a non-sampling based visualization.